Make configure install locally rather than globally.
By default the configure script wants SDL2 to be located in
/usr/local/include/SDL2. We pass extra commandline args to change that
to the current package's location. This helps when building `tsdl` which
will run `pkg-config --cflags sdl2` which looks into `sdl2.pc` which
is generated when configure is run.

Also for some reason a file called libSDL2main.a is generated but not
libSDL2.a, which is the file that the linker looks for when it sees the
flag `-lSDL2`. We change that flag to be `-lSDLmain` instead.
This could be a cross platform issue when building on windows/linux
(we need to try it).